qr/requirements.txt
certifi==2019.3.9
chardet==3.0.4
enum34==1.1.6
idna==2.8
Pillow==6.0.0
qrcode==6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vk-api==11.4.0
